Added spells to turn-based combat
There have also been a significant number of system changes, including the conversion of everything to use Tile objects, which simplifies all grid interactions. Also note that I have separated the Combat UI from the GameController object, meaning it no longer has to be treated as a Prefab and can be left in the Combat scene and modified there. This was done to make testing and UI changes easier.